Add new intratime functions and tests
Improve rest of modules

+++ test/unit/conftest.py
@@ -3,7 +3,11 @@ import os
 
 from intratime_slack_bot.lib.test_utils import TEST_FILE
 
+# ----------------------------------------------------------------------------------------------------------------------
+
+
 @pytest.fixture
 def remove_test_file(request):
     yield
-    os.remove(TEST_FILE)
+    if os.path.exists(TEST_FILE):
+        os.remove(TEST_FILE)
